+++ packages-config/eslint-config/config/config-es.js
@@ -1,0 +1,11 @@
+import eslintJs from '@eslint/js';
+
+import rulesEs from '../rules/rules-es.js';
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  eslintJs.configs.recommended,
+  {
+    rules: rulesEs
+  }
+];