+++ packages-config/eslint-config/config/config-node.js
@@ -1,0 +1,6 @@
+import pluginNode from 'eslint-plugin-n';
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  pluginNode.configs['flat/recommended-script']
+];