+++ packages-config/eslint-config/config/config-react-hooks.js
@@ -3,11 +3,9 @@ import {
 } from 'eslint/config';
 import pluginReactHooks from 'eslint-plugin-react-hooks';
 
-import rulesReactHooks from './rules/rules-react-hooks.js';
+import rulesReactHooks from '../rules/rules-react-hooks.js';
 
-/**
- * @type {import('eslint').Linter.Config[]}
- */
+/** @type {import('eslint').Linter.Config[]} */
 export default defineConfig([
   pluginReactHooks.configs.flat.recommended,
   {