+++ packages-config/eslint-config/config/config-ts-type-unaware.js
@@ -1,12 +1,12 @@
 import eslintTs from 'typescript-eslint';
 
-import rules from './rules/rules-ts-type-unaware.js';
+import rulesTsTypeUnaware from '../rules/rules-ts-type-unaware.js';
 
 /** @type {import('eslint').Linter.Config[]} */
 export default eslintTs.config(
     ...eslintTs.configs.stylistic, // eslint-disable-line import/no-named-as-default-member
     ...eslintTs.configs.strict, // eslint-disable-line import/no-named-as-default-member
     {
-      rules
+      rules: rulesTsTypeUnaware
     }
 );