+++ packages-config/eslint-config/config/config-vue.js
@@ -1,0 +1,11 @@
+import pluginVue from 'eslint-plugin-vue';
+
+import rulesVue from '../rules/rules-vue.js';
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  ...pluginVue.configs['flat/recommended'],
+  {
+    rules: rulesVue
+  }
+];