+++ packages-config/eslint-config/index.js
@@ -1,14 +1,16 @@
 import globals from 'globals';
 
 // 相对引用必须 .js 否则 ERR_MODULE_NOT_FOUND
-import configEs from './config-es.js';
-import configTs from './config-ts-type-aware.js';
-import configReact from './config-react.js';
-import configReactHooks from './config-react-hooks.js';
-import configImport from './config-import.js';
-import configStylistic from './config-stylistic.js';
+import configEs from './config/config-es.js';
+import configTs from './config/config-ts-type-aware.js';
+import configReact from './config/config-react.js';
+import configReactHooks from './config/config-react-hooks.js';
+import configImport from './config/config-import.js';
+import configStylistic from './config/config-stylistic.js';
 
 /**
+ * Web 端 React 项目可用
+ *
  * 为何分两个 `index`？这个跟 typescript-eslint 有关，见 https://typescript-eslint.io/getting-started/typed-linting#shared-configurations
  *
  * ```diff