+++ packages-config/eslint-config/rules/rules-vue.js
@@ -1,0 +1,22 @@
+/**
+ * https://eslint.vuejs.org/rules/
+ *
+ * @type {import('eslint').Linter.RulesRecord}
+ */
+export default {
+  'vue/multi-word-component-names': 'off',
+  'vue/singleline-html-element-content-newline': 'off',
+  'vue/max-attributes-per-line': 'off',
+  'vue/block-order': ['error', {
+    order: ['template', 'script', 'style']
+  }],
+  'vue/html-self-closing': ['error', {
+    html: {
+      void: 'always',
+      normal: 'always',
+      component: 'always'
+    },
+    svg: 'always',
+    math: 'always'
+  }]
+};